# docs/swagger.yaml
definitions:
...
  dto.BindTokenDto:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  dto.CreatePlaceDto:
    properties:
      address:
        maxLength: 300
        minLength: 1
        type: string
      city:
        maxLength: 300
        minLength: 1
        type: string
      description:
        maxLength: 300
        minLength: 1
        type: string
      name:
        maxLength: 300
        minLength: 1
        type: string
    required:
    - address
    - city
    - description
    - name
    type: object
...
  dto.CreateUserDto:
    properties:
      email:
        maxLength: 64
        minLength: 6
        type: string
      password:
        maxLength: 64
        minLength: 6
        type: string
      telegram_username:
        maxLength: 32
        minLength: 4
        type: string
      username:
        maxLength: 32
        minLength: 6
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.GetUserDto:
    properties:
      created_at:
        type: string
      email:
        type: string
      is_admin:
        type: boolean
      telegram_username:
        maxLength: 32
        minLength: 4
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  dto.PlaceDto:
    properties:
      address:
        maxLength: 300
        minLength: 1
        type: string
      city:
        maxLength: 300
        minLength: 1
        type: string
      description:
        maxLength: 300
        minLength: 1
        type: string
      id:
        type: integer
      name:
        maxLength: 300
        minLength: 1
        type: string
    required:
    - address
    - city
    - description
    - name
    type: object
...
  dto.UpdateUserDto:
    properties:
      password:
        type: string
      telegram_username:
        maxLength: 32
        minLength: 4
        type: string
      updated_at:
        description: sets in service automatically
        type: string
      username:
        maxLength: 32
        minLength: 6
        type: string
    type: object
  dto.UserWithRatingDto:
    properties:
      created_at:
        type: string
      email:
        type: string
      is_admin:
        type: boolean
      rating:
        type: number
      telegram_username:
        maxLength: 32
        minLength: 4
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  exceptions.Error_:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /places:
    get:
      consumes:
      - application/json
      description: Возвращает все места
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PlaceDto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/exceptions.Error_'
      security:
      - BearerAuth: []
      summary: Получить все места
      tags:
      - places
    post:
      consumes:
      - application/json
      description: Создает новое место (доступно только администраторам)
      parameters:
      - description: Данные нового места
        in: body
        name: place
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePlaceDto'
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PlaceDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/exceptions.Error_'
      security:
      - BearerAuth: []
      summary: Создать место
      tags:
      - places
  /places/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет место по ID
      parameters:
      - description: ID места
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/exceptions.Error_'
      security:
      - BearerAuth: []
      summary: Удалить место
      tags:
      - places
...
  /users:
    post:
      consumes:
      - application/json
      description: Creates new user and returns it
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GetUserDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/exceptions.Error_'
      summary: Создание пользователя
      tags:
      - user
  /users/{username}:
    get:
      description: Returns user profile by username (requires JWT in "Bearer" header)
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserWithRatingDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/exceptions.Error_'
      security:
      - BearerAuth: []
      summary: Получить профиль пользователя (по нику)
      tags:
      - user
...
  /users/bind_token:
    post:
      consumes:
      - application/json
      description: Биндит к юзеру токен мобильного устройства
      parameters:
      - description: Token data
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/dto.BindTokenDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/exceptions.Error_'
      summary: Бинд токена для уведомлений
      tags:
      - user
  /users/me:
    get:
      description: Returns user profile (requires JWT in "Bearer" header)
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserWithRatingDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/exceptions.Error_'
      security:
      - BearerAuth: []
      summary: Получить профиль пользователя (свой)
      tags:
      - user
    patch:
      description: Updates user profile and returns it (requires JWT in "Bearer" header)
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDto'
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetUserDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.Error_'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/exceptions.Error_'
      security:
      - BearerAuth: []
      summary: Обновить профиль.
      tags:
      - user
swagger: "2.0"